Moved string extensions to separate file. Improved syntax
